# docs/openapi.yaml
openapi: 3.0.3
info:
  title: BitByBit Habits API
  description: |
    API REST de BitByBit, un habit tracker que recompensa completar tareas con sats reales via Lightning Network.

    **Caso de uso MVP**: "Mesada 2.0" donde sponsors (padres/tutores) crean habitos con recompensa en sats
    y kids los completan para ganar.

    ## Autenticacion

    La API usa autenticacion basada en cookies de sesion (`session`). Para autenticarse:

    1. Registrar un usuario con `POST /api/auth/register`
    2. Iniciar sesion con `POST /api/auth/login`
    3. La cookie `session` se setea automaticamente en la respuesta

    Todos los endpoints (excepto auth) requieren una sesion activa.

    ## Formato de respuesta

    Todas las respuestas siguen el formato `ApiResponse`:
    ```json
    {
      "success": true,
      "data": { ... }
    }
    ```
    En caso de error:
    ```json
    {
      "success": false,
      "error": "Descripcion del error"
    }
    ```
  version: 1.0.0
  contact:
    name: BitByBit Team
servers:
  - url: /api
    description: Servidor de la API
...
paths:
...
  /auth/session:
    get:
      tags:
        - Auth
      summary: Obtener sesion actual
      description: Devuelve los datos de la sesion del usuario autenticado.
      operationId: getSession
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Sesion activa encontrada
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/AuthSession"
              example:
                success: true
                data:
                  user_id: "550e8400-e29b-41d4-a716-446655440000"
                  email: "padre@ejemplo.com"
                  username: "papajuan"
                  display_name: "Juan"
                  role: "sponsor"
                  locale: "es"
        "401":
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"
              example:
                success: false
                error: "No autenticado"
  /habits:
    get:
      tags:
        - Habits
      summary: Listar habitos del usuario
      description: |
        Devuelve los habitos relevantes para el usuario autenticado.

        - **Kid**: habitos asignados a el dentro de sus familias.
        - **Sponsor**: habitos creados por el dentro de sus familias.

        Cada habito incluye el campo `completed_today` indicando si ya fue completado hoy.
      operationId: getHabits
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Lista de habitos obtenida exitosamente
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/HabitWithCompletion"
              example:
                success: true
                data:
                  - id: "660e8400-e29b-41d4-a716-446655440001"
                    family_id: "770e8400-e29b-41d4-a716-446655440002"
                    created_by: "550e8400-e29b-41d4-a716-446655440000"
                    assigned_to: "880e8400-e29b-41d4-a716-446655440003"
                    name: "Tender la cama"
                    description: "Tender la cama todas las mananas"
                    color: "#4CAF7D"
                    icon: "bed"
                    sat_reward: 50
                    schedule_type: "daily"
                    verification_type: "sponsor_approval"
                    active: true
                    completed_today: false
                    created_at: "2026-03-01T10:00:00.000Z"
                    updated_at: "2026-03-01T10:00:00.000Z"
        "401":
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"
              example:
                success: false
                error: "No autenticado"
        "500":
          description: Error al obtener los habitos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"
              example:
                success: false
                error: "Error al obtener los habitos"
...
  /families:
    get:
      tags:
        - Families
      summary: Listar familias del usuario
      description: |
        Devuelve todas las familias a las que pertenece el usuario autenticado,
        incluyendo los miembros de cada familia con sus datos basicos.
      operationId: getFamilies
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Lista de familias obtenida exitosamente
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/FamilyWithMembers"
              example:
                success: true
                data:
                  - id: "770e8400-e29b-41d4-a716-446655440002"
                    name: "Familia Rodriguez"
                    invite_code: "ABC123"
                    created_by: "550e8400-e29b-41d4-a716-446655440000"
                    created_at: "2026-03-01T10:00:00.000Z"
                    members:
                      - id: "aaa00000-0000-0000-0000-000000000001"
                        role: "sponsor"
                        joined_at: "2026-03-01T10:00:00.000Z"
                        user_id: "550e8400-e29b-41d4-a716-446655440000"
                        display_name: "Juan"
                        username: "papajuan"
                        avatar_url: null
                      - id: "aaa00000-0000-0000-0000-000000000002"
                        role: "kid"
                        joined_at: "2026-03-02T10:00:00.000Z"
                        user_id: "880e8400-e29b-41d4-a716-446655440003"
                        display_name: "Lucia"
                        username: "luciarodriguez"
                        avatar_url: null
        "401":
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"
              example:
                success: false
                error: "No autenticado"
        "500":
          description: Error al obtener las familias
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"
              example:
                success: false
                error: "Error al obtener las familias"

    post:
      tags:
        - Families
      summary: Crear una nueva familia
      description: |
        Crea una nueva familia y agrega al usuario como miembro sponsor.
        Solo usuarios con rol `sponsor` pueden crear familias.
        Se genera automaticamente un codigo de invitacion de 6 caracteres.
      operationId: createFamily
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFamilyRequest"
            example:
              name: "Familia Rodriguez"
      responses:
        "201":
          description: Familia creada exitosamente
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Family"
              example:
                success: true
                data:
                  id: "770e8400-e29b-41d4-a716-446655440002"
                  name: "Familia Rodriguez"
                  invite_code: "ABC123"
                  created_by: "550e8400-e29b-41d4-a716-446655440000"
                  created_at: "2026-03-01T10:00:00.000Z"
        "400":
          description: Nombre de familia faltante
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"
              example:
                success: false
                error: "El nombre de la familia es obligatorio"
        "401":
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"
              example:
                success: false
                error: "No autenticado"
        "403":
          description: Solo los sponsors pueden crear familias
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"
              example:
                success: false
                error: "Solo los sponsors pueden crear familias"
        "500":
          description: Error al crear la familia
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"
              example:
                success: false
                error: "Error al crear la familia"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
      description: Cookie de sesion httpOnly seteada por el endpoint de login
  schemas:
...
    CreateFamilyRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: Nombre de la familia
...
    ApiResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          description: Indica si la operacion fue exitosa
        data:
          description: Datos de la respuesta (presente cuando success es true)
        error:
          type: string
          description: Mensaje de error (presente cuando success es false)
    ApiResponseError:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          enum:
            - false
        error:
          type: string
          description: Mensaje de error descriptivo
...
    AuthSession:
      type: object
      description: Datos de la sesion del usuario autenticado
      properties:
        user_id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        username:
          type: string
        display_name:
          type: string
        role:
          type: string
          enum:
            - sponsor
            - kid
        locale:
          type: string
          enum:
            - es
            - en
    Habit:
      type: object
      description: Habito creado por un sponsor y asignado a un kid
      properties:
        id:
          type: string
          format: uuid
        family_id:
          type: string
          format: uuid
          description: ID de la familia a la que pertenece
        created_by:
          type: string
          format: uuid
          description: ID del sponsor que lo creo
        assigned_to:
          type: string
          format: uuid
          description: ID del kid al que esta asignado
        name:
          type: string
          description: Nombre del habito
        description:
          type: string
          nullable: true
          description: Descripcion opcional del habito
        color:
          type: string
          description: Color del habito (hex)
          example: "#4CAF7D"
        icon:
          type: string
          nullable: true
          description: Icono opcional del habito
        sat_reward:
          type: integer
          description: Recompensa en sats al completar
        schedule_type:
          type: string
          enum:
            - daily
            - specific_days
            - times_per_week
          description: Tipo de programacion del habito
        schedule_days:
          type: array
          items:
            type: integer
          nullable: true
          description: Dias especificos (0=domingo, 6=sabado). Solo para schedule_type specific_days.
        schedule_times_per_week:
          type: integer
          nullable: true
          description: Veces por semana. Solo para schedule_type times_per_week.
        verification_type:
          type: string
          enum:
            - sponsor_approval
            - self_verify
            - bot_verify
          description: Como se verifica la completacion
        active:
          type: boolean
          description: Si el habito esta activo
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    HabitWithCompletion:
      allOf:
        - $ref: "#/components/schemas/Habit"
        - type: object
          properties:
            completed_today:
              type: boolean
              description: Si el habito ya fue completado hoy
...
    Family:
      type: object
      description: Grupo familiar
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          description: Nombre de la familia
        invite_code:
          type: string
          description: Codigo de invitacion de 6 caracteres
        created_by:
          type: string
          format: uuid
          description: ID del sponsor que la creo
        created_at:
          type: string
          format: date-time
...
    FamilyMemberWithUser:
      type: object
      description: Miembro de familia con datos del usuario
      properties:
        id:
          type: string
          format: uuid
        role:
          type: string
          enum:
            - sponsor
            - kid
        joined_at:
          type: string
          format: date-time
        user_id:
          type: string
          format: uuid
        display_name:
          type: string
        username:
          type: string
        avatar_url:
          type: string
          format: uri
          nullable: true
    FamilyWithMembers:
      allOf:
        - $ref: "#/components/schemas/Family"
        - type: object
          properties:
            members:
              type: array
              items:
                $ref: "#/components/schemas/FamilyMemberWithUser"
...